feat: 添加 Pinia 状态管理
- 安装 Pinia 依赖
- 创建 task store 集中管理任务状态
- 将轮询逻辑从 composable 迁移到 store
- 重构 App.vue 使用 store 管理状态
- 删除不再需要的 useTaskPolling composable
