+++ frontend/package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "marked": "^15.0.4",
+    "pinia": "^3.0.4",
     "vue": "^3.5.13"
   },
   "devDependencies": {
